read magick

+++ udp.go
@@ -80,10 +80,10 @@ func (u *UDP) Write(b []byte) (n int, err error) {
 		return
 	}
 	n += 2
-	if u.Length >= 8 {
-		u.Data = make([]byte, u.Length-8)
+	if u.Length >= uint16(8) {
+		u.Data = make([]byte, u.Length-uint16(8))
 	}
-	if u.Length == 0 {
+	if u.Length == uint16(0) {
 		u.Data = make([]byte, buf.Len())
 	}
 	m, err := io.ReadFull(buf, u.Data)
